Ad Db Contracts as example. Remove deprecated version from docker-compose.
