+++ docker-compose.yml
@@ -1,4 +1,3 @@
-version: '3.0'
 services:
     php-static-resolver-bundle:
         image: pimcore/pimcore:php8.2-debug-dev
